Release 1.0.4 (#59)
* Bug fix

* signup bug fix

* user invite

* docs: add local setup instructions, cleanup files (#10)

* Update README.md (#11)

* Update README.md

* Update README.md

* Update README.md

* Bugfix auth (#12)

* Update README.md

* Update favicon (#17)

* Release 1.0.3 (#8)

* Bug fix

* signup bug fix

* user invite

* favicon

* Adding pagination to traces (#16)

* Adding pagination to traces

* installing scroller

* query function bug

* fix

* Bug fix pagination

---------


* Dylan/s3en 2060 adding more frontend pagination (#18)

* Adding pagination to traces

* installing scroller

* bug fixes for trace pagination

* adding delete prompt api

* adding pagination to evaluate

* adding pagination to evals

* adding pagination to prompset, bug fixes

* adding .env

* fix

---------


* updating prompt dialog (#19)

* Show API key warning message (#20)

* Show API key warning message

* Update instructions

* Fix null dereference checks (#21)

* adding check for missing table (#22)

* adding check for missing table

* adding docs button

* Check empty table

---------


* Check length of data (#24)

* flicker issue (#26)

* Flicker issue (#27)

* flicker issue

* Fixes

* Fix (#30)

* Fix header (#32)

* updating gitignore

* Rohit/s3 en 2064 fix api key not showing (#29)

* fix api key bug

* final fixes

* fix scroll bar issue

* adjust readme

* Pagination bug

* Pagination bug (#37)

* removing form reset on edit (#39)

* removing form reset on edit

* bug fixes:

* fix

* fix page number bug

---------


* Dylan/s3en 2079 thumbs up duplication bug (#40)

* refetch/duplicate bug fix

* Prevent stale data in forms

* fix

* linting

* fixing duplication bug

* removing console

* Fix eval

* Fix pagination for promp eval

* Fix promptsets pagination

---------


* Bug fixes for Traces Pagination (#42)

* Pagination bug

* Fix pagination for traces

* Bug fix

* Fix invalid model (#43)

* Pagination bug

* Bug fix

* Bug fixes (#45)

* Pagination bug

* Bug fix

* fix pricing

* fix the pii function

* Bugfixes

* adding pagination to dataset

* Trace page improvements (#49)

* Pagination bug

* Bug fix

* Fix scrolling pagination

* Filters

* Add model

* utc - local time switch

* Bugfixes and Evaluation Dashboard (#50)

* Pagination bug

* Bug fix

* minor fix

* Evaluations dashboard refresh

* Move prompts as a tab

* UI improvements

* round values

* Add migration script

* Minor bug fixes (#51)

* Pagination bug

* Bug fix

* Bug fix

* Disable LLM view for framework and vector db requests

* Minor bug fixes (#52)

* Pagination bug

* Bug fix

* Bug fix

* Disable LLM view for framework and vector db requests

* adding projects page loading

* deleting

* fix

* adding perplexity cost calculator:

* Dylan/s3en 2127 loading states for all pages (#57)

* loading for data set

* adding loading page for promptset

* dataset loading

* prompts loading

* prompts loading

* metrics loading, chart loading

* evaluations loading

* refactor traces

* trace row loading

* more loading

* Rename loading to skeleton

* Minor fix

---------


* Bugfixes and cleanups (#58)

* Pagination bug

* Bug fix

* Make /projects endpoint response cleaner

* Better error handling for projects page

* Cleanup metrics page

* Cleanup metrics and better error handling

* better error handling for traces

* Error handling for evaluations page

* Error handle evaluations

* Error handling for datasets

* Minor fixes

* hover fixes

* Minor fix

* minor fix

---------
