+++ app/(protected)/project/[project_id]/datasets/dataset/[dataset_id]/page.tsx
@@ -1,6 +1,7 @@
 "use client";
 
 import { CreateData } from "@/components/project/dataset/create-data";
+import DatasetRowSkeleton from "@/components/project/dataset/dataset-row-skeleton";
 import { EditData } from "@/components/project/dataset/edit-data";
 import { Spinner } from "@/components/shared/spinner";
 import { Button } from "@/components/ui/button";
@@ -12,6 +13,7 @@ import { useParams } from "next/navigation";
 import { useState } from "react";
 import { useBottomScrollListener } from "react-bottom-scroll-listener";
 import { useQuery } from "react-query";
+import { toast } from "sonner";
 
 export default function Dataset() {
   const dataset_id = useParams()?.dataset_id as string;
@@ -36,6 +38,10 @@ export default function Dataset() {
       const response = await fetch(
         `/api/dataset?dataset_id=${dataset_id}&page=${page}&pageSize=${PAGE_SIZE}`
       );
+      if (!response.ok) {
+        const error = await response.json();
+        throw new Error(error?.message || "Failed to fetch dataset");
+      }
       const result = await response.json();
       return result;
     },
@@ -66,10 +72,16 @@ export default function Dataset() {
       }
       setShowLoader(false);
     },
+    onError: (error) => {
+      setShowLoader(false);
+      toast.error("Failed to fetch dataset", {
+        description: error instanceof Error ? error.message : String(error),
+      });
+    },
   });
 
   if (fetchDataset.isLoading || !fetchDataset.data || !currentData) {
-    return <div>Loading...</div>;
+    return <PageSkeleton />;
   } else {
     return (
       <div className="w-full py-6 px-6 flex flex-col gap-4">
@@ -129,4 +141,33 @@ export default function Dataset() {
       </div>
     );
   }
+}
+
+function PageSkeleton() {
+  return (
+    <div className="w-full py-6 px-6 flex flex-col gap-4">
+      <div className="flex gap-4 items-center w-fit">
+        <Button
+          disabled={true}
+          variant="secondary"
+          onClick={() => window.history.back()}
+        >
+          <ChevronLeft className="mr-1" />
+          Back
+        </Button>
+        <CreateData disabled={true} />
+      </div>
+      <div className="flex flex-col gap-3 rounded-md border border-muted max-h-screen overflow-y-scroll">
+        <div className="grid grid-cols-5 items-center justify-stretch gap-3 py-3 px-4 bg-muted">
+          <p className="text-xs font-medium">Created at</p>
+          <p className="text-xs font-medium">Input</p>
+          <p className="text-xs font-medium">Output</p>
+          <p className="text-xs font-medium text-end">Note</p>
+        </div>
+        {Array.from({ length: 5 }).map((_, index) => (
+          <DatasetRowSkeleton key={index} />
+        ))}
+      </div>
+    </div>
+  );
 }
