+++ app/(protected)/project/[project_id]/prompts/page-client.tsx
@@ -5,15 +5,17 @@ import { Spinner } from "@/components/shared/spinner";
 import { Button } from "@/components/ui/button";
 import { Checkbox } from "@/components/ui/checkbox";
 import { Separator } from "@/components/ui/separator";
+import { Skeleton } from "@/components/ui/skeleton";
 import { PAGE_SIZE } from "@/lib/constants";
 import { extractPromptFromLlmInputs } from "@/lib/utils";
 import { CheckCircledIcon } from "@radix-ui/react-icons";
-import { ChevronDown, ChevronRight } from "lucide-react";
+import { ChevronDown, ChevronRight, RabbitIcon } from "lucide-react";
 import { useParams } from "next/navigation";
 import { useState } from "react";
 import { useBottomScrollListener } from "react-bottom-scroll-listener";
 import Markdown from "react-markdown";
 import { useQuery } from "react-query";
+import { toast } from "sonner";
 
 interface CheckedData {
   value: string;
@@ -21,7 +23,7 @@ interface CheckedData {
 }
 
 export default function PageClient({ email }: { email: string }) {
-  const project_id = useParams()?.project_id as string;
+  const projectId = useParams()?.project_id as string;
   const [selectedData, setSelectedData] = useState<CheckedData[]>([]);
   const [page, setPage] = useState<number>(1);
   const [totalPages, setTotalPages] = useState<number>(1);
@@ -37,11 +39,15 @@ export default function PageClient({ email }: { email: string }) {
   };
 
   const fetchPrompts = useQuery({
-    queryKey: ["fetch-prompts-query"],
+    queryKey: [`fetch-prompts-${projectId}-query`],
     queryFn: async () => {
       const response = await fetch(
-        `/api/prompt?projectId=${project_id}&page=${page}&pageSize=${PAGE_SIZE}`
+        `/api/prompt?projectId=${projectId}&page=${page}&pageSize=${PAGE_SIZE}`
       );
+      if (!response.ok) {
+        const error = await response.json();
+        throw new Error(error?.message || "Failed to fetch prompts");
+      }
       const result = await response.json();
       return result;
     },
@@ -72,6 +78,13 @@ export default function PageClient({ email }: { email: string }) {
       }
       setShowLoader(false);
     },
+    onError: (error) => {
+      setCurrentData([]);
+      setShowLoader(false);
+      toast.error("Failed to fetch prompts", {
+        description: error instanceof Error ? error.message : String(error),
+      });
+    },
   });
 
   useBottomScrollListener(() => {
@@ -84,8 +97,17 @@ export default function PageClient({ email }: { email: string }) {
     }
   });
 
-  if (fetchPrompts.isLoading || !fetchPrompts.data || !currentData) {
-    return <div>Loading...</div>;
+  if (fetchPrompts.isLoading) {
+    return <PageLoading />;
+  } else if (fetchPrompts.error) {
+    return (
+      <div className="md:px-52 px-12 py-12 flex flex-col items-center justify-center">
+        <RabbitIcon size={80} />
+        <p className="text-lg font-semibold">
+          An error occurred while fetching prompts. Please try again later.
+        </p>
+      </div>
+    );
   } else {
     // Deduplicate prompts
     const seenPrompts: string[] = [];
@@ -109,7 +131,7 @@ export default function PageClient({ email }: { email: string }) {
     return (
       <div className="w-full py-6 px-6 flex flex-col gap-4">
         <div className="w-fit">
-          <AddtoPromptset projectId={project_id} selectedData={selectedData} />
+          <AddtoPromptset projectId={projectId} selectedData={selectedData} />
         </div>
         <p className="text-sm font-semibold text-black bg-yellow-300 px-2 p-1 rounded-md">
           These prompts are automatically captured from your traces. The
@@ -171,9 +193,18 @@ const PromptRow = ({
     queryKey: [`fetch-promptdata-query-${prompt.span_id}`],
     queryFn: async () => {
       const response = await fetch(`/api/promptdata?spanId=${prompt.span_id}`);
+      if (!response.ok) {
+        const error = await response.json();
+        throw new Error(error?.message || "Failed to fetch prompt data");
+      }
       const result = await response.json();
       setAddedToPromptset(result.data.length > 0);
       return result;
+    },
+    onError: (error) => {
+      toast.error("Failed to fetch prompt data", {
+        description: error instanceof Error ? error.message : String(error),
+      });
     },
   });
 
@@ -186,8 +217,13 @@ const PromptRow = ({
     queryKey: [`fetch-evaluation-query-${prompt.span_id}`],
     queryFn: async () => {
       const response = await fetch(`/api/evaluation?prompt=${promptContent}`);
-      const result = await response.json();
 
+      if (!response.ok) {
+        const error = await response.json();
+        throw new Error(error?.message || "Failed to fetch evaluation");
+      }
+
+      const result = await response.json();
       // calculate accuracy
       let score = 0;
       result.evaluations.forEach((e: any) => {
@@ -198,6 +234,11 @@ const PromptRow = ({
       setAccuracy((score * 100) / result.evaluations.length);
 
       return result;
+    },
+    onError: (error) => {
+      toast.error("Failed to fetch evaluation", {
+        description: error instanceof Error ? error.message : String(error),
+      });
     },
   });
 
@@ -280,3 +321,60 @@ const PromptRow = ({
     </div>
   );
 };
+
+function PageLoading() {
+  return (
+    <div className="w-full py-6 px-6 flex flex-col gap-4">
+      <div className="w-fit">
+        <AddtoPromptset disabled={true} />
+      </div>
+      <p className="text-sm font-semibold text-black bg-yellow-300 px-2 p-1 rounded-md">
+        These prompts are automatically captured from your traces. The accuracy
+        of these prompts are calculated based on the evaluation done in the
+        evaluate tab.
+      </p>
+      <div className="flex flex-col gap-3 rounded-md border border-muted max-h-screen overflow-y-scroll">
+        <div className="grid grid-cols-6 items-center justify-stretch gap-3 py-3 px-4 bg-muted">
+          <p className="text-xs font-medium">LLM Vendor</p>
+          <p className="text-xs font-medium text-left">Model</p>
+          <p className="text-xs font-medium text-left">Interactions</p>
+          <p className="text-xs text-left font-medium">Prompt</p>
+          <p className="text-xs font-medium text-center">Accuracy</p>
+          <p className="text-xs font-medium">Added to Dataset</p>
+        </div>
+        {Array.from({ length: 3 }).map((_, index) => (
+          <PromptRowSkeleton key={index} />
+        ))}
+      </div>
+    </div>
+  );
+}
+
+function PromptRowSkeleton() {
+  return (
+    <div className="flex flex-col gap-3">
+      <div className="grid grid-cols-6 justify-stretch items-center py-1 px-4 w-full cursor-pointer">
+        <div
+          className="flex flex-row items-center gap-2"
+          onClick={(e) => e.stopPropagation()}
+        >
+          <Skeleton className="w-full h-6" />
+        </div>
+        <div className="text-xs text-muted-foreground text-left font-semibold">
+          <Skeleton className="w-full h-6" />
+        </div>
+        <div className="text-xs text-left font-semibold">
+          <Skeleton className="w-full h-6" />
+        </div>
+        <div className="text-xs text-left h-10 truncate overflow-y-scroll font-semibold">
+          <Skeleton className="w-full h-6" />
+        </div>
+        <div className="text-xs text-center font-semibold">
+          <Skeleton className="w-full h-6" />
+        </div>
+        <Skeleton className="w-full h-6" />
+      </div>
+      <Separator orientation="horizontal" />
+    </div>
+  );
+}
