+++ app/(protected)/projects/page-client.tsx
@@ -2,6 +2,7 @@
 
 import { Create } from "@/components/project/create";
 import { Edit } from "@/components/project/edit";
+import CardSkeleton from "@/components/shared/card-skeleton";
 import {
   Card,
   CardContent,
@@ -12,47 +13,87 @@ import {
 import { cn } from "@/lib/utils";
 import { Project } from "@prisma/client";
 import { ArrowTopRightIcon } from "@radix-ui/react-icons";
+import { RabbitIcon } from "lucide-react";
 import Link from "next/link";
 import { useQuery } from "react-query";
+import { toast } from "sonner";
 
 export default function PageClient({ email }: { email: string }) {
-  const fetchProjects = useQuery({
+  const {
+    data: projects,
+    isLoading: projectsLoading,
+    error: projectsError,
+  } = useQuery({
     queryKey: ["fetch-projects-query"],
     queryFn: async () => {
       const response = await fetch(`/api/projects?email=${email}`);
+      if (!response.ok) {
+        const error = await response.json();
+        throw new Error(error?.message || "Failed to fetch projects");
+      }
       const result = await response.json();
       return result;
     },
+    onError: (error) => {
+      toast.error("Failed to fetch projects", {
+        description: error instanceof Error ? error.message : String(error),
+      });
+    },
   });
 
-  const fetchUser = useQuery({
+  const {
+    data: user,
+    isLoading: userLoading,
+    error: userError,
+  } = useQuery({
     queryKey: ["fetch-user-query"],
     queryFn: async () => {
       const response = await fetch(`/api/user?email=${email}`);
+      if (!response.ok) {
+        const error = await response.json();
+        throw new Error(error?.message || "Failed to fetch user");
+      }
       const result = await response.json();
       return result;
     },
+    onError: (error) => {
+      toast.error("Failed to fetch user", {
+        description: error instanceof Error ? error.message : String(error),
+      });
+    },
   });
 
-  if (
-    fetchProjects.isLoading ||
-    !fetchProjects.data ||
-    fetchUser.isLoading ||
-    !fetchUser.data
-  ) {
-    return <div>Loading...</div>;
+  if (projectsLoading || userLoading) {
+    return <PageSkeleton />;
+  }
+
+  if (projectsError || userError) {
+    return (
+      <div className="w-full flex flex-col">
+        <div className="md:px-52 px-12 py-12 flex justify-between bg-muted">
+          <h1 className="text-3xl font-semibold">Projects</h1>
+          <Create teamId={user?.data?.Team?.id} />
+        </div>
+        <div className="md:px-52 px-12 py-12 flex flex-col items-center justify-center">
+          <RabbitIcon size={80} />
+          <p className="text-lg text-destructive font-semibold">
+            An error occurred while fetching data. Please try again later.
+          </p>
+        </div>
+      </div>
+    );
   }
 
   return (
     <div className="w-full flex flex-col">
       <div className="md:px-52 px-12 py-12 flex justify-between bg-muted">
         <h1 className="text-3xl font-semibold">Projects</h1>
-        <Create teamId={fetchUser.data.data.Team.id} />
+        <Create teamId={user?.data?.Team?.id} />
       </div>
       <div
         className={cn(
           "md:px-52 px-12 py-12 flex md:flex-row flex-col gap-2 items-center",
-          fetchProjects.data.data.projects.length === 0
+          projects?.projects?.length === 0
             ? "md:items-center"
             : "md:items-start"
         )}
@@ -60,29 +101,27 @@ export default function PageClient({ email }: { email: string }) {
         <div
           className={cn(
             "flex w-full gap-12 flex-wrap",
-            fetchProjects.data.data.projects.length === 0
+            projects?.projects?.length === 0
               ? "flex-col items-center"
               : "md:flex-row flex-wrap flex-col md:items-start items-center"
           )}
         >
-          {fetchProjects.data.data.projects.length === 0 && (
+          {projects?.projects?.length === 0 && (
             <div className="flex flex-col gap-2 items-center">
               <p className="text-2xl text-muted-foreground">Welcome!</p>
               <p className="text-lg text-muted-foreground mb-2">
                 Create a new project to get started
               </p>
-              <Create teamId={fetchUser.data.data.Team.id} />
+              <Create teamId={user?.data?.Team?.id} />
             </div>
           )}
-          {fetchProjects.data.data.projects.map(
-            (project: Project, i: number) => (
-              <ProjectCard
-                key={i}
-                project={project}
-                teamId={fetchUser.data.data.Team.id}
-              />
-            )
-          )}
+          {projects?.projects?.map((project: Project, i: number) => (
+            <ProjectCard
+              key={i}
+              project={project}
+              teamId={user?.data?.Team?.id}
+            />
+          ))}
         </div>
       </div>
     </div>
@@ -98,26 +137,32 @@ function ProjectCard({
   project: Project;
   teamId: string;
 }) {
-  const fetchProjecStats = useQuery({
+  const { data: projectStats, isLoading: projectStatsLoading } = useQuery({
     queryKey: [`fetch-project-stats-${project.id}`],
     queryFn: async () => {
       const response = await fetch(
         `/api/stats/project?projectId=${project.id}`
       );
+      if (!response.ok) {
+        const error = await response.json();
+        throw new Error(error?.message || "Failed to fetch project stats");
+      }
       const result = await response.json();
       return result;
     },
+    onError: (error) => {
+      toast.error("Failed to fetch project stats", {
+        description: error instanceof Error ? error.message : String(error),
+      });
+    },
   });
 
-  if (fetchProjecStats.isLoading || !fetchProjecStats.data) {
-    return <div>Loading...</div>;
-  }
   return (
     <div className="relative" key={key}>
       <div className="flex items-center flex-row gap-2 absolute top-2 right-2 z-10">
-        {!fetchProjecStats.isLoading &&
-          fetchProjecStats.data &&
-          fetchProjecStats.data.totalSpans === 0 && (
+        {!projectStatsLoading &&
+          projectStats &&
+          projectStats?.totalSpans === 0 && (
             <Link
               href={`/project/${project.id}/traces`}
               className="cursor-pointer flex flex-row gap-2 h-8 text bg-orange-300 hover:bg-orange-400 dark:bg-orange-600 dark:hover:bg-orange-700 p-2 rounded-md"
@@ -130,9 +175,7 @@ function ProjectCard({
       </div>
       <Link
         href={
-          !fetchProjecStats.isLoading &&
-          fetchProjecStats.data &&
-          fetchProjecStats.data.totalSpans === 0
+          !projectStatsLoading && projectStats && projectStats?.totalSpans === 0
             ? `/project/${project.id}/traces`
             : `/project/${project.id}/metrics`
         }
@@ -147,25 +190,25 @@ function ProjectCard({
             </CardDescription>
           </CardHeader>
           <CardContent>
-            {!fetchProjecStats.isLoading && fetchProjecStats.data && (
+            {!projectStatsLoading && projectStats && (
               <div className="flex flex-row justify-between">
                 <div className="flex flex-col gap-2">
                   <div className="flex flex-row gap-1 items-center">
                     <p className="text-sm text-muted-foreground">Traces</p>
                     <p className="text-sm font-semibold">
-                      {fetchProjecStats.data.totalTraces}
+                      {projectStats?.totalTraces || 0}
                     </p>
                   </div>
                   <div className="flex flex-row gap-1 items-center">
                     <p className="text-sm text-muted-foreground">Spans</p>
                     <p className="text-sm font-semibold">
-                      {fetchProjecStats.data.totalSpans}
+                      {projectStats?.totalSpans || 0}
                     </p>
                   </div>
                   <div className="flex flex-row gap-1 items-center">
                     <p className="text-sm text-muted-foreground">Evaluations</p>
                     <p className="text-sm font-semibold">
-                      {fetchProjecStats.data.totalEvaluations}
+                      {projectStats?.totalEvaluations || 0}
                     </p>
                   </div>
                 </div>
@@ -173,13 +216,13 @@ function ProjectCard({
                   <div className="flex flex-row gap-1 items-center">
                     <p className="text-sm text-muted-foreground">Datasets</p>
                     <p className="text-sm font-semibold">
-                      {fetchProjecStats.data.totalDatasets}
+                      {projectStats?.totalDatasets || 0}
                     </p>
                   </div>
                   <div className="flex flex-row gap-1 items-center">
                     <p className="text-sm text-muted-foreground">Prompt sets</p>
                     <p className="text-sm font-semibold">
-                      {fetchProjecStats.data.totalPromptsets}
+                      {projectStats?.totalPromptsets || 0}
                     </p>
                   </div>
                 </div>
@@ -188,6 +231,32 @@ function ProjectCard({
           </CardContent>
         </Card>
       </Link>
+    </div>
+  );
+}
+
+export function PageSkeleton() {
+  return (
+    <div className="w-full flex flex-col">
+      <div className="md:px-52 px-12 py-12 flex justify-between bg-muted">
+        <h1 className="text-3xl font-semibold">Projects</h1>
+        <Create disabled={true} />
+      </div>
+      <div
+        className={cn(
+          "md:px-52 px-12 py-12 flex md:flex-row flex-col gap-2 items-center md:items-start"
+        )}
+      >
+        <div
+          className={cn(
+            "flex w-full gap-12 flex-wrap md:flex-row flex-col md:items-start items-center"
+          )}
+        >
+          {Array.from({ length: 3 }).map((_, index) => (
+            <CardSkeleton key={index} />
+          ))}
+        </div>
+      </div>
     </div>
   );
 }
