+++ components/charts/token-chart.tsx
@@ -2,21 +2,36 @@
 
 import { BarChart } from "@tremor/react";
 import { useQuery } from "react-query";
+import { toast } from "sonner";
+import SmallChartLoading from "./small-chart-skeleton";
 
 export function TokenChart({ projectId }: { projectId: string }) {
-  const fetchMetricsUsageToken = useQuery({
-    queryKey: ["fetch-metrics-usage-token-query"],
+  const {
+    data: tokenUsage,
+    isLoading: tokenUsageLoading,
+    error: tokenUsageError,
+  } = useQuery({
+    queryKey: [`fetch-metrics-usage-token-${projectId}-query`],
     queryFn: async () => {
       const response = await fetch(
         `/api/metrics/usage/token?projectId=${projectId}`
       );
+      if (!response.ok) {
+        const error = await response.json();
+        throw new Error(error?.message || "Failed to fetch token usage");
+      }
       const result = await response.json();
       return result;
     },
+    onError: (error) => {
+      toast.error("Failed to fetch token usage", {
+        description: error instanceof Error ? error.message : String(error),
+      });
+    },
   });
 
-  if (fetchMetricsUsageToken.isLoading || !fetchMetricsUsageToken.data) {
-    return <div>Loading...</div>;
+  if (tokenUsageLoading || tokenUsageError) {
+    return <SmallChartLoading />;
   } else {
     return (
       <>
@@ -24,25 +39,23 @@ export function TokenChart({ projectId }: { projectId: string }) {
           <div className="flex flex-row gap-4 h-12">
             <div className="flex flex-col gap-1">
               <p className="text-sm font-semibold text-start">
-                Total Input Tokens:{" "}
-                {fetchMetricsUsageToken.data.totalInputTokens || 0}
+                Total Input Tokens: {tokenUsage?.totalInputTokens || 0}
               </p>
               <p className="text-sm font-semibold text-start">
-                Total Output Tokens:{" "}
-                {fetchMetricsUsageToken.data.totalOutputTokens || 0}
+                Total Output Tokens: {tokenUsage?.totalOutputTokens || 0}
               </p>
             </div>
             <p className="text-sm font-semibold text-start">
-              Total Tokens: {fetchMetricsUsageToken.data.totalTokens || 0}
+              Total Tokens: {tokenUsage?.totalTokens || 0}
             </p>
           </div>
           <BarChart
             className="mt-2 h-72"
-            data={fetchMetricsUsageToken.data?.usage?.map((data: any) => ({
-              date: data.date,
-              "Total Tokens": parseInt(data.totalTokens),
-              "Input Tokens": parseInt(data.inputTokens),
-              "Output Tokens": parseInt(data.outputTokens),
+            data={tokenUsage?.usage?.map((data: any) => ({
+              date: data?.date,
+              "Total Tokens": parseInt(data?.totalTokens),
+              "Input Tokens": parseInt(data?.inputTokens),
+              "Output Tokens": parseInt(data?.outputTokens),
             }))}
             index="date"
             categories={["Total Tokens", "Input Tokens", "Output Tokens"]}
@@ -62,19 +75,32 @@ export function TokenChart({ projectId }: { projectId: string }) {
 }
 
 export function CostChart({ projectId }: { projectId: string }) {
-  const fetchMetricsUsageCost = useQuery({
-    queryKey: ["fetch-metrics-usage-cost-query"],
+  const {
+    data: costUsage,
+    isLoading: costUsageLoading,
+    error: costUsageError,
+  } = useQuery({
+    queryKey: [`fetch-metrics-usage-cost-${projectId}-query`],
     queryFn: async () => {
       const response = await fetch(
         `/api/metrics/usage/cost?projectId=${projectId}`
       );
+      if (!response.ok) {
+        const error = await response.json();
+        throw new Error(error?.message || "Failed to fetch cost usage");
+      }
       const result = await response.json();
       return result;
     },
+    onError: (error) => {
+      toast.error("Failed to fetch cost usage", {
+        description: error instanceof Error ? error.message : String(error),
+      });
+    },
   });
 
-  if (fetchMetricsUsageCost.isLoading || !fetchMetricsUsageCost.data) {
-    return <div>Loading...</div>;
+  if (costUsageLoading || costUsageError) {
+    return <SmallChartLoading />;
   } else {
     return (
       <>
@@ -82,25 +108,23 @@ export function CostChart({ projectId }: { projectId: string }) {
           <div className="flex flex-row gap-4 h-12">
             <div className="flex flex-col gap-1">
               <p className="text-sm font-semibold text-start">
-                Input Tokens Cost: $
-                {fetchMetricsUsageCost.data?.input?.toFixed(6) || 0}
+                Input Tokens Cost: ${costUsage?.input?.toFixed(6) || 0}
               </p>
               <p className="text-sm font-semibold text-start">
-                Output Tokens Cost: $
-                {fetchMetricsUsageCost.data?.output?.toFixed(6) || 0}
+                Output Tokens Cost: ${costUsage?.output?.toFixed(6) || 0}
               </p>
             </div>
             <p className="text-sm font-semibold text-start">
-              Total Cost: ${fetchMetricsUsageCost.data?.total?.toFixed(6) || 0}
+              Total Cost: ${costUsage?.total?.toFixed(6) || 0}
             </p>
           </div>
           <BarChart
             className="mt-2 h-72"
-            data={fetchMetricsUsageCost.data?.cost?.map((data: any) => ({
+            data={costUsage?.cost?.map((data: any) => ({
               date: data?.date,
-              "Total Cost": data?.total?.toFixed(6),
-              "Input Tokens Cost": data?.input?.toFixed(6),
-              "Output Tokens Cost": data?.output?.toFixed(6),
+              "Total Cost": data?.total?.toFixed(6) || 0,
+              "Input Tokens Cost": data?.input?.toFixed(6) || 0,
+              "Output Tokens Cost": data?.output?.toFixed(6) || 0,
             }))}
             index="date"
             categories={[
