+++ components/shared/hover-cell.tsx
@@ -17,8 +17,8 @@ export function HoverCell({
       <HoverCardTrigger asChild>
         <button className={className}>{value}</button>
       </HoverCardTrigger>
-      <HoverCardContent className="w-full whitespace-pre-wrap">
-        <Markdown className={"break-all"}>{value}</Markdown>
+      <HoverCardContent className="w-[40rem] max-h-[20rem] p-4 overflow-y-scroll whitespace-pre-wrap text-sm">
+        <Markdown className="break-all">{value}</Markdown>
       </HoverCardContent>
     </HoverCard>
   );
